Extract blast facet filtering into separate util file.
