+++ src/tools/blast/utils/__tests__/blastFacetDataUtils.spec.ts
@@ -1,0 +1,52 @@
+// const t = filterBlastDatum(
+//   { score: 300, evalue: 0, identity: 80 },
+//   { score: [400, 600], identity: [70, 90] },
+//   'score'
+// );
+
+// console.log("blast test:", t);
+
+// const row = {
+//   score: 100,
+//   identity: 10,
+//   evalue: 1,
+// };
+
+// const test_1 = filterBlastDatum(row);
+
+// console.log("test 1:", test_1);
+
+// const test_2 = filterBlastDatum(row,
+//   { score: [85, 150]},
+//   'score'
+// );
+
+// console.log("test 2:", test_2);
+
+// const test_3 = filterBlastDatum(row,
+//   { score: [200, 300]},
+//   'score'
+// );
+
+// console.log("test 3:", test_3);
+
+// const test_4 = filterBlastDatum(row,
+//   { score: [50, 150], identity: [20, 30]},
+//   'score'
+// );
+
+// console.log("test 4:", test_4);
+
+// const test_5 = filterBlastDatum(row,
+//   { score: [200, 300], identity: [5, 20]},
+//   'identity'
+// );
+
+// console.log("test 5:", test_5);
+
+// const test_6 = filterBlastDatum(row,
+//   { score: [10, 200], identity: [95, 100]},
+//   'identity'
+// );
+
+// console.log("test 6:", test_6);
